Exception handling - delete

+++ src/main/java/br/com/educandoweb/curso/model/servicos/UsuarioServico.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import br.com.educandoweb.curso.model.entidades.Usuario;
 import br.com.educandoweb.curso.model.repositorios.UsuarioRepositorio;
+import br.com.educandoweb.curso.model.servicos.exceptions.DatabaseException;
 import br.com.educandoweb.curso.model.servicos.exceptions.ResourceNotFoundException;
 
 @Service
@@ -21,7 +24,7 @@ public class UsuarioServico {
 
 	public Usuario findById(Long id) {
 		Optional<Usuario> obj = repositorio.findById(id);
-		return obj.orElseThrow(()-> new ResourceNotFoundException(id));
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 
 	public Usuario insert(Usuario obj) {
@@ -29,9 +32,18 @@ public class UsuarioServico {
 	}
 
 	public void delete(Long id) {
-		repositorio.deleteById(id);
+		try {
+			if (repositorio.existsById(id)) {
+				repositorio.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
+
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public Usuario update(Long id, Usuario obj) {
 		Usuario entity = repositorio.getReferenceById(id);
 		updateData(entity, obj);
@@ -41,6 +53,6 @@ public class UsuarioServico {
 	private void updateData(Usuario entity, Usuario obj) {
 		entity.setNome(obj.getNome());
 		entity.setEmail(obj.getEmail());
-		}
+	}
 
 }
